プレイヤー弾画像変更、プレイヤー死亡時に動きを停止させる処理の追加、プレイヤー死亡時にレベルアップでHPを増加させた時に死亡用画像のまま続行出来てしまう不具合を修正

+++ competition/Objects/Character/Player/PlayerStats.h
@@ -19,7 +19,7 @@ struct PlayerStats
 	int Shield_count = 3;
 	int player_level = 1;
 	int player_hp_max = 6;
-	int player_shot_hitrange_up = 0;
+	float player_shot_hitrange_up = 0.0f;
 	int drone_count = 0;
 	bool threeway_flag = false;
 	bool drone_flag = false;
